app/src/test/java/com/example/youtube_philiplackner_flow/MainViewModelTest.kt
package com.example.youtube_philiplackner_flow

import app.cash.turbine.test
import com.google.common.truth.Truth.assertThat
import kotlinx.coroutines.ExperimentalCoroutinesApi
import kotlinx.coroutines.launch
import kotlinx.coroutines.runBlocking
import org.junit.Before
import org.junit.Test

class MainViewModelTest {

    private lateinit var viewModel: MainViewModel
    private lateinit var testDispatchers: TestDispatchers

    @Before
    fun setUp() {
        testDispatchers = TestDispatchers()
        viewModel = MainViewModel(testDispatchers)
    }

    @OptIn(ExperimentalCoroutinesApi::class)
    @Test
    fun `countDownFlow, properly counts down from 10 to 0`() = runBlocking {
        viewModel.countdownFlow.test {

            for (i in 10 downTo 0) {
                testDispatchers.testDispatcher.scheduler.advanceTimeBy(1000L)
                val emission = awaitItem()
                assertThat(emission).isEqualTo(i)

            }
            cancelAndConsumeRemainingEvents()
        }
    }

    @Test
    fun `squareNumber, number properly squared`() = runBlocking {
        val job = launch {
            viewModel.sharedFlow.test {

                val emission = awaitItem()

                assertThat(emission).isEqualTo(9)
                cancelAndConsumeRemainingEvents()
            }
        }
        viewModel.squareNumber(3)
        job.join()
        job.cancel()

    }

}